refactor(watch): split the recovery path out of store_bolt.go
The previous commit pushed store_bolt.go to 834 lines, past the repository's
800-line gate. I saw that failure and pushed anyway: the command ran `make dod`,
`git commit` and `git push` with no dependency between them, so a red gate
printed its error and the chain carried on. That is the defect class this whole
release has been about, committed on the check whose job is to prevent it.

Split rather than allowlisted. The gate offers both, and an allowlist entry buys
the same green light while teaching nothing. This code has needed five
revisions, each one a way for recovery to destroy the data it exists to save,
so it earns its own file and a header that says as much:

  quarantining on any load failure          -- a 5s lock wait destroys history
  quarantining on any schema failure        -- a downgrade destroys history
  quarantining on a missing version key     -- corruption of one key does
  measuring emptiness with Stats()          -- BucketN counts the bucket itself,
                                               so recovery never fired at all
  renaming without re-checking under lock   -- a writer publishing into the same
                                               file gets quarantined

None was caught by the test suite or the gates. All five came from adversarial
review, by three different vendors, each finding what the others had passed.

Pure code movement: the sentinel, the two path helpers, the quarantine, the
unpublished re-check, neverPublished, and the atomic first publish. No logic
changed, and the eight boundary tests that pin them are unmoved and still pass.

V: make dod exit 0, captured as an exit code rather than read off the log --
which is how the previous failure got past me.
